chore: move pinpoint logic into pinpoint-helper and reuse region from meta
