+++ packages/amplify-category-notifications/package.json
@@ -29,11 +29,14 @@
     "@aws-amplify/amplify-cli-core": "4.0.1",
     "@aws-amplify/amplify-environment-parameters": "1.4.1",
     "@aws-amplify/amplify-prompts": "2.6.6",
+    "@aws-amplify/amplify-provider-awscloudformation": "8.0.3",
+    "aws-sdk": "^2.1233.0",
     "chalk": "^4.1.1",
     "fs-extra": "^8.1.0",
     "lodash": "^4.17.21",
     "ora": "^4.0.3",
-    "promise-sequential": "^1.1.1"
+    "promise-sequential": "^1.1.1",
+    "proxy-agent": "^5.0.0"
   },
   "jest": {
     "testURL": "http://localhost",
